refactor(RankingList): 랭킹 아이템을 별도 파일로 분리하여 재사용 가능하게 함
